refactor/move message related functions to message file
