+++ message.go
@@ -84,3 +84,51 @@ type Field struct {
 	Value  string `json:"value"`
 	Inline bool   `json:"inline,omitempty"`
 }
+
+// divideMessages breaks message into multiple messages depending on embed total
+// character count and number of embeds.
+func divideMessages(messages []Message) (msgs []Message) {
+	for _, msg := range messages {
+		dividedEmbeds := divideEmbeds(msg)
+		// Create message for every embed chunk.
+		for i, embeds := range dividedEmbeds {
+			// First message contains content from original message.
+			var content string
+			var files []*File
+			if i == 0 {
+				content = msg.Content
+				files = msg.Files
+			}
+			msgs = append(msgs, Message{Username: msg.Username, Embeds: embeds, Content: content, Files: files})
+		}
+	}
+	return msgs
+}
+
+func divideEmbeds(msg Message) (dividedEmbeds [][]Embed) {
+	var total int
+	var startIndex int
+	for i, e := range msg.Embeds {
+		count := countEmbed(e)
+		total += count
+		// i will be included in next chunk.
+		if total > EmbedTotalLimit || i-startIndex == MessageEmbedNumLimit {
+			dividedEmbeds = append(dividedEmbeds, msg.Embeds[startIndex:i])
+			startIndex = i
+			// Set current count to initial total of next message.
+			total = count
+		}
+	}
+	// Add the last chunk.
+	dividedEmbeds = append(dividedEmbeds, msg.Embeds[startIndex:])
+	return
+}
+
+func countEmbed(e Embed) int {
+	total := len(e.Title) + len(e.Description) + len(e.Author.Name) + len(e.Footer.Text)
+	for _, field := range e.Fields {
+		total += len(field.Name)
+		total += len(field.Value)
+	}
+	return total
+}
